+++ src/main/java/org/opentripplanner/routing/api/request/preference/TransferPreferences.java
@@ -4,10 +4,12 @@ import static java.util.Objects.requireNonNull;
 import static org.opentripplanner.framework.lang.DoubleUtils.doubleEquals;
 
 import java.io.Serializable;
+import java.time.Duration;
 import java.util.Objects;
 import java.util.function.Consumer;
 import org.opentripplanner.framework.model.Cost;
 import org.opentripplanner.framework.model.Units;
+import org.opentripplanner.framework.time.DurationUtils;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.routing.algorithm.transferoptimization.api.TransferOptimizationParameters;
 
@@ -23,7 +25,7 @@ public final class TransferPreferences implements Serializable {
   private static final int MAX_NUMBER_OF_TRANSFERS = 30;
 
   private final Cost cost;
-  private final int slack;
+  private final Duration slack;
   private final double waitReluctance;
   private final int maxTransfers;
   private final int maxAdditionalTransfers;
@@ -32,7 +34,7 @@ public final class TransferPreferences implements Serializable {
 
   private TransferPreferences() {
     this.cost = Cost.ZERO;
-    this.slack = 120;
+    this.slack = Duration.ofMinutes(2);
     this.waitReluctance = 1.0;
     this.maxTransfers = 12;
     this.maxAdditionalTransfers = 5;
@@ -42,7 +44,7 @@ public final class TransferPreferences implements Serializable {
 
   private TransferPreferences(Builder builder) {
     this.cost = builder.cost;
-    this.slack = Units.duration(builder.slack);
+    this.slack = DurationUtils.requireNonNegative(builder.slack);
     this.waitReluctance = Units.reluctance(builder.waitReluctance);
     this.maxTransfers = Units.count(builder.maxTransfers, MAX_NUMBER_OF_TRANSFERS);
     this.maxAdditionalTransfers =
@@ -78,18 +80,18 @@ public final class TransferPreferences implements Serializable {
   }
 
   /**
-   * A global minimum transfer time (in seconds) that specifies the minimum amount of time that must
+   * A global minimum transfer time that specifies the minimum amount of time that must
    * pass between exiting one transit vehicle and boarding another. This time is in addition to time
    * it might take to walk between transit stops, the {@link TransitPreferences#alightSlack()}, and the {@link
-   * TransitPreferences#boardSlack()}. This time should also be overridden by specific transfer timing information in
-   * transfers.txt
+   * TransitPreferences#boardSlack()}.
+   * This time can also be overridden by specific transfer timing information in transfers.txt
    * <p>
-   * This only apply to transfer between two trips, it does not apply when boarding the first
+   * This only applies to transfer between two trips, it does not apply when boarding the first
    * transit.
    * <p>
-   * Unit is seconds. Default value is 2 minutes.
+   * Default value is 2 minutes.
    */
-  public int slack() {
+  public Duration slack() {
     return slack;
   }
 
@@ -156,7 +158,7 @@ public final class TransferPreferences implements Serializable {
     TransferPreferences that = (TransferPreferences) o;
     return (
       cost.equals(that.cost) &&
-      slack == that.slack &&
+      slack.equals(that.slack) &&
       doubleEquals(that.waitReluctance, waitReluctance) &&
       maxTransfers == that.maxTransfers &&
       maxAdditionalTransfers == that.maxAdditionalTransfers &&
@@ -183,7 +185,7 @@ public final class TransferPreferences implements Serializable {
     return ToStringBuilder
       .of(TransferPreferences.class)
       .addObj("cost", cost, DEFAULT.cost)
-      .addNum("slack", slack, DEFAULT.slack)
+      .addDuration("slack", slack, DEFAULT.slack)
       .addNum("waitReluctance", waitReluctance, DEFAULT.waitReluctance)
       .addNum("maxTransfers", maxTransfers, DEFAULT.maxTransfers)
       .addNum("maxAdditionalTransfers", maxAdditionalTransfers, DEFAULT.maxAdditionalTransfers)
@@ -196,7 +198,7 @@ public final class TransferPreferences implements Serializable {
 
     private final TransferPreferences original;
     private Cost cost;
-    private int slack;
+    private Duration slack;
     private Integer maxTransfers;
     private Integer maxAdditionalTransfers;
     private double waitReluctance;
@@ -223,7 +225,11 @@ public final class TransferPreferences implements Serializable {
       return this;
     }
 
-    public Builder withSlack(int slack) {
+    public Builder withSlackSec(Number seconds) {
+      return withSlack(Duration.ofSeconds(seconds.longValue()));
+    }
+
+    public Builder withSlack(Duration slack) {
       this.slack = slack;
       return this;
     }