+++ src/test/java/org/opentripplanner/routing/api/request/framework/LinearFunctionSerializationTest.java
@@ -36,6 +36,8 @@ class LinearFunctionSerializationTest {
            3h + 5.111 t   ||       3h  |   5.1
             7m + 10.1 x   ||       7m  |  10.0
           PT7s + 10.1 x   ||       7s  |  10.0
+          0.1 + 10.1 x   ||       0s  |  10.0
+          0.5 + 10.1 x   ||       1s  |  10.0
         """
     );
   }
@@ -53,7 +55,7 @@ class LinearFunctionSerializationTest {
   }
 
   @Test
-  void parseEmtpy() {
+  void parseEmpty() {
     assertEquals(Optional.empty(), LinearFunctionSerialization.parse(null, fail()));
     assertEquals(Optional.empty(), LinearFunctionSerialization.parse("", fail()));
     assertEquals(Optional.empty(), LinearFunctionSerialization.parse(" \r\n", fail()));
@@ -75,6 +77,15 @@ class LinearFunctionSerializationTest {
       () -> LinearFunctionSerialization.parse("foo", fail())
     );
     assertEquals("Unable to parse function: 'foo'", ex.getMessage());
+  }
+
+  @Test
+  void parseIllegalDuration() {
+    var ex = assertThrows(
+      IllegalArgumentException.class,
+      () -> LinearFunctionSerialization.parse("600ss + 1.3 t", fail())
+    );
+    assertEquals("Unable to parse duration: '600ss'", ex.getMessage());
   }
 
   private static BiFunction<Duration, Double, ?> fail() {